BaseResponse class
