+++ lib/fondy.rb
@@ -5,6 +5,7 @@ require 'fondy/version'
 require 'fondy/errors'
 require 'fondy/client'
 require 'fondy/request'
+require 'fondy/base_response'
 require 'fondy/response'
 require 'fondy/transaction_list_response'
 require 'fondy/signature'
